+++ modules/registry-authorizer/versions.tf
@@ -1,3 +1,17 @@
 terraform {
   required_version = ">= 1.5.0"
+  required_providers {
+    aws = {
+      source  = "hashicorp/aws"
+      version = "~> 5.5.0"
+    }
+    archive = {
+      source  = "hashicorp/archive"
+      version = "2.4.0"
+    }
+    external = {
+      source  = "hashicorp/external"
+      version = "2.3.1"
+    }
+  }
 }