+++ modules/registry-store/main.tf
@@ -25,11 +25,24 @@ resource "aws_dynamodb_table" "modules" {
     name = "Version"
     type = "S"
   }
+  #tfsec:ignore:aws-dynamodb-table-customer-key less cost 
+  server_side_encryption { # default alias/aws/dynamodb
+    enabled     = true
+  }
+
+  point_in_time_recovery {
+    enabled = var.enable_point_in_time_recovery
+  }
 
   tags = merge(var.tags, { Name : local.dynamodb_table_name })
 }
 
 
+# tfsec:ignore:aws-s3-block-public-acls see ressource aws_s3_bucket_acl.bucket
+# tfsec:ignore:aws-s3-block-public-policy see aws_s3_bucket_public_access_block.bucket
+# tfsec:ignore:aws-s3-enable-bucket-encryption see aws_s3_bucket_server_side_encryption_configuration.bucket
+# tfsec:ignore:aws-s3-encryption-customer-key see aws_s3_bucket_server_side_encryption_configuration.bucket
+# tfsec:ignore:aws-s3-enable-bucket-logging access logging is done with api gateway
 resource "aws_s3_bucket" "bucket" {
   bucket = local.bucket_name
   tags   = merge(var.tags, { Name : local.bucket_name })
@@ -52,5 +65,23 @@ resource "aws_s3_bucket_versioning" "bucket" {
   bucket = aws_s3_bucket.bucket.id
   versioning_configuration {
     status = "Enabled"
+  }
+}
+
+resource "aws_s3_bucket_public_access_block" "default" {
+  bucket                  = aws_s3_bucket.bucket.id
+  block_public_acls       = var.public_access.block_public_acls
+  ignore_public_acls      = var.public_access.ignore_public_acls
+  block_public_policy     = var.public_access.block_public_policy
+  restrict_public_buckets = var.public_access.restrict_public_buckets
+}
+
+#tfsec:ignore:aws-s3-encryption-customer-key
+resource "aws_s3_bucket_server_side_encryption_configuration" "bucket" {
+  bucket = aws_s3_bucket.bucket.id
+  rule {
+    apply_server_side_encryption_by_default {
+      sse_algorithm = "AES256"
+    }
   }
 }